backend/app/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
pydantic
google-genai
httpx
